Removed ITSM-CIAttributeCollection.sopm
When installing the package you will get "ITSM-CIAttributeCollection 6.0.2 - Package not correctly deployed! Please reinstall the package." because "ITSM-CIAttributeCollection.sopm - File is not installed!"
This file is not necessary when the opm package is build as far as I know. After removing the file from the opm file list and reinstalling it on OTRS 6.0.8 the error was gone.